:bug: Fix foreign key in vehicle component table

+++ database/migrations/base_structure/api/starcitizen/vehicle/vehicle/2020_09_18_124902_create_vehicle_component_table.php
@@ -24,7 +24,7 @@ class CreateVehicleComponentTable extends Migration
         Schema::create(
             'vehicle_component',
             function (Blueprint $table) {
-                $table->unsignedBigInteger('vehicle_id');
+                $table->unsignedInteger('vehicle_id');
                 $table->unsignedBigInteger('component_id');
 
                 $table->foreign('vehicle_id')
